pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reading_list"
version = "2.3.0"
description = "Reading List Tracker"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
requires-python = ">=3.8"
dependencies = [
    "flask>=2.0.0",
    "flask-wtf>=1.0.0",
    "sqlalchemy>=1.4.0",
    "alembic>=1.7.0",
    "pandas>=1.3.0",
    "xlsxwriter>=3.0.0",
    "openpyxl>=3.0.0",
    "tabulate>=0.9.0",
    "termcolor>=2.0.0",
    "rich>=10.0.0",
    "python-dotenv>=0.19.0",
    "requests>=2.28.0",
    "aiohttp>=3.8.0",
    "fastapi>=0.68.0",
    "Pillow>=10.0.0",
    "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "mypy>=0.900",
    "pylint>=2.0.0"
]

[project.scripts]
reading-list = "reading_list.cli.main:main"
reading-list-covers = "reading_list.cli.covers:main"
reading-api = "reading_list.api.app:app"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["reading_list"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = ["."]

[tool.mypy]
python_version = "2.3.0"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.python.compile-bytecode]
enabled = false
